requirements.txt
amqp==2.6.0
asgiref==3.2.9
beautifulsoup4==4.9.1
billiard==3.6.3.0
bs4==0.0.1
celery==4.4.5
certifi==2020.4.5.2
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
dj-rest-auth==1.0.9
Django==3.0.7
django-admin-autocomplete-filter==0.5
django-admin-rangefilter==0.6.1
django-admin-sortable==2.2.3
django-admin-sortable2==0.7.6
django-admin-tools==0.9.0
django-allauth==0.42.0
django-autocomplete-light==3.5.1
django-ckeditor==5.9.0
django-colorfield==0.3.1
django-cors-headers==3.4.0
django-crispy-forms==1.9.1
django-debug-toolbar==2.2
django-decouple==2.1
django-filebrowser-no-grappelli==3.8.0
django-filer==1.7.1
django-froala-editor==3.1.0
django-grappelli==2.14.2
django-import-export==2.2.0
django-js-asset==1.2.2
django-markdownx==3.0.1
django-modeltranslation==0.15
django-mptt==0.11.0
django-nested-admin==3.3.2
django-polymorphic==2.1.2
django-rest-framework==0.1.0
django-rosetta==0.9.4
django-suit==0.2.26
django-summernote==0.8.11.6
django-templated-mail==1.1.1
django-tinymce4-lite==1.8.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
easy-thumbnails==2.7
et-xmlfile==1.0.1
future==0.18.2
idna==2.9
importlib-metadata==1.6.1
jdcal==1.4.1
jsmin==2.2.2
kombu==4.6.10
Markdown==3.2.2
MarkupPy==1.14
numpy==1.18.5
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
pandas==1.0.4
Pillow==7.1.2
pkg-resources==0.0.0
polib==1.1.0
PyJWT==1.7.1
python-dateutil==2.8.1
python-monkey-business==1.0.0
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tablib==2.0.0
transliterate==1.10.2
Unidecode==1.1.1
urllib3==1.25.9
vine==1.3.0
xlrd==1.2.0
XlsxWriter==1.2.9
xlwt==1.3.0
zipp==3.1.0
